+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Change Log
 - Developer/alpha release (**STILL NOT FOR PRODUCTION USE!**)
 - Features and significant changes:
   + Implemented man pages for all Syncany commands: `man (sy|sy-up|..)` #87
+- Bugfixes:
+  + Fixed issue with renaming to folder with quotes on Windows #124 
+  + Fix uploading large files, i.e. action file handler using same connection
+    as upload; #140
 
 ### Syncany 0.1.5-alpha (Date: 12 Jun 2014)